Updated Data Classes & Setting up path for initializing via a database
Edited the data classes so that its hierarchy is more organized and better fits the app architecture.

As the creation of MainActivity, the model will be initialized with data from a database (represented for now as a Mock Database).

If there is no data to be got from the database, the app will initialize a "Guest User" and "Guest Budget".